// source/types/radioChannels.ts
export enum RadioChannel {
	DIRECTOR_PLAY = 'director:play',
	DIRECTOR_PAUSE = 'director:pause',
	DIRECTOR_SCENE_CHANGE = 'director:scenechange',
	DIRECTOR_SCENE_EVENT = 'director:sceneevent',
	VIDEO_ENDED = 'video:ended',
	VIDEO_DURATION_CHANGED = 'video:durationchange',
	VIDEO_PAUSED = 'video:paused',
	VIDEO_PLAYING = 'video:playing',
	VIDEO_SEEKED = 'video:seeked',
	VIDEO_SEEKING = 'video:seeking',
	VIDEO_TIMEUPDATE = 'video:timeupdate'
}